backend/pyproject.toml
[project]
name = "rivalry-research"
version = "0.1.0"
description = "A Python library for analyzing rivalrous relationships using Wikidata's knowledge graph"
readme = "README.md"
authors = [
    { name = "David Rojas", email = "lempira@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "pydantic-ai>=1.18.0",
    "pydantic-settings>=2.0.0",
    "beautifulsoup4>=4.14.2",
    "google-genai>=1.50.0",
    "logfire>=0.50.0",
    "scholarly>=1.7.11",
    "pymupdf>=1.26.6",
    "arxiv>=2.3.1",
    "typer>=0.15.1",
    "pillow>=11.0.0",
    "rich>=13.0.0",
]

[project.scripts]
rivalry = "rivalry_research.cli.main:app"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
